Extract Query runner tests from production modules
